CLI: first atempt at terminals.
Not fully working.
After launching a term, the script doesn't wait for the job to end. Thus
ret is always 0 and runs the next terminal with a screen job, which
blocks the port and prevents first terminal from uploading the sketch.

As a workaround, we can wait for a char to be typed. Then the user has
to wait to end the upload and reach for both terminals at the same time.

Ugly, but ...

Gonna try with term functions.